moved to pow.rs
